# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Checkout & Rewards",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/carts/{cart_id}": {
      "get": {
        "tags": [
          "Carts"
        ],
        "summary": "View a cart",
        "description": "Lines show the current unit price, current stock and an `in_stock` flag, so a UI can warn before\ncheckout. Prices aren't locked in until checkout.",
        "operationId": "get_cart_carts__cart_id__get",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Cart ID returned by `POST /carts`.",
              "examples": [
                "b685d68c87cb40ffbb3c88af5c8cdfad"
              ],
              "title": "Cart Id"
            },
            "description": "Cart ID returned by `POST /carts`."
          }
        ],
        "responses": {
          "200": {
            "description": "The cart, priced at **current** product prices.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                },
                "examples": {
                  "open": {
                    "summary": "Open cart",
                    "value": {
                      "id": "b685d68c87cb40ffbb3c88af5c8cdfad",
                      "status": "open",
                      "items": [
                        {
                          "product_id": "p_cable",
                          "name": "USB-C Cable",
                          "unit_price_minor": 999,
                          "quantity": 2,
                          "line_total_minor": 1998,
                          "available_inventory": 500,
                          "in_stock": true
                        },
                        {
                          "product_id": "p_keyboard",
                          "name": "Mechanical Keyboard",
                          "unit_price_minor": 8999,
                          "quantity": 1,
                          "line_total_minor": 8999,
                          "available_inventory": 50,
                          "in_stock": true
                        }
                      ],
                      "subtotal_minor": 10997,
                      "currency": "INR"
                    }
                  },
                  "out_of_stock": {
                    "summary": "Stock dropped after the item was added",
                    "value": {
                      "id": "b685d68c87cb40ffbb3c88af5c8cdfad",
                      "status": "open",
                      "subtotal_minor": 44997,
                      "currency": "INR",
                      "items": [
                        {
                          "product_id": "p_lamp",
                          "name": "Limited Edition Desk Lamp",
                          "unit_price_minor": 14999,
                          "quantity": 3,
                          "line_total_minor": 44997,
                          "available_inventory": 2,
                          "in_stock": false
                        }
                      ]
                    }
                  },
                  "checked_out": {
                    "summary": "Cart after checkout",
                    "value": {
                      "id": "b685d68c87cb40ffbb3c88af5c8cdfad",
                      "status": "checked_out",
                      "order_id": "bb5df4bff6d1471cb976cb781c95a385",
                      "items": [
                        {
                          "product_id": "p_cable",
                          "name": "USB-C Cable",
                          "unit_price_minor": 999,
                          "quantity": 2,
                          "line_total_minor": 1998,
                          "available_inventory": 500,
                          "in_stock": true
                        },
                        {
                          "product_id": "p_keyboard",
                          "name": "Mechanical Keyboard",
                          "unit_price_minor": 8999,
                          "quantity": 1,
                          "line_total_minor": 8999,
                          "available_inventory": 50,
                          "in_stock": true
                        }
                      ],
                      "subtotal_minor": 10997,
                      "currency": "INR"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "examples": {
                  "CART_NOT_FOUND": {
                    "summary": "Unknown cart ID",
                    "value": {
                      "error": {
                        "code": "CART_NOT_FOUND",
                        "message": "Cart 0d4c6b1e9f2a4e7c8b3d5a6f1e2c9b70 does not exist",
                        "details": {}
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/carts/{cart_id}/items": {
      "post": {
        "tags": [
          "Carts"
        ],
        "summary": "Add an item",
        "description": "Adds a product that isn't in the cart yet. If it's already there, this returns `409 ITEM_ALREADY_IN_CART`\ninstead of adding more, so a retried request can't double the quantity. Use `PUT` to change a quantity.\n\nThe stock check here is advisory: carts don't reserve stock, and checkout re-checks it.",
        "operationId": "add_item_carts__cart_id__items_post",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Cart ID returned by `POST /carts`.",
              "examples": [
                "b685d68c87cb40ffbb3c88af5c8cdfad"
              ],
              "title": "Cart Id"
            },
            "description": "Cart ID returned by `POST /carts`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItem"
              },
              "examples": {
                "keyboard": {
                  "summary": "Add one keyboard",
                  "value": {
                    "product_id": "p_keyboard",
                    "quantity": 1
                  }
                },
                "limited": {
                  "summary": "Add the limited-stock lamp",
                  "value": {
                    "product_id": "p_lamp",
                    "quantity": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The updated cart.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                },
                "examples": {
                  "added": {
                    "summary": "Item added",
                    "value": {
                      "id": "b685d68c87cb40ffbb3c88af5c8cdfad",
                      "status": "open",
                      "items": [
                        {
                          "product_id": "p_cable",
                          "name": "USB-C Cable",
                          "unit_price_minor": 999,
                          "quantity": 2,
                          "line_total_minor": 1998,
                          "available_inventory": 500,
                          "in_stock": true
                        },
                        {
                          "product_id": "p_keyboard",
                          "name": "Mechanical Keyboard",
                          "unit_price_minor": 8999,
                          "quantity": 1,
                          "line_total_minor": 8999,
                          "available_inventory": 50,
                          "in_stock": true
                        }
                      ],
                      "subtotal_minor": 10997,
                      "currency": "INR"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "examples": {
                  "CART_NOT_FOUND": {
                    "summary": "Unknown cart ID",
                    "value": {
                      "error": {
                        "code": "CART_NOT_FOUND",
                        "message": "Cart 0d4c6b1e9f2a4e7c8b3d5a6f1e2c9b70 does not exist",
                        "details": {}
                      }
                    }
                  },
                  "PRODUCT_NOT_FOUND": {
                    "summary": "Unknown product ID",
                    "value": {
                      "error": {
                        "code": "PRODUCT_NOT_FOUND",
                        "message": "Product p_speaker does not exist",
                        "details": {}
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with current state",
            "content": {
              "application/json": {
                "examples": {
                  "ITEM_ALREADY_IN_CART": {
                    "summary": "Product already in the cart (use PUT)",
                    "value": {
                      "error": {
                        "code": "ITEM_ALREADY_IN_CART",
                        "message": "Product already in cart; use PUT to change its quantity",
                        "details": {}
                      }
                    }
                  },
                  "INSUFFICIENT_INVENTORY_ADD": {
                    "summary": "More than is in stock right now",
                    "value": {
                      "error": {
                        "code": "INSUFFICIENT_INVENTORY",
                        "message": "Not enough inventory for requested quantity",
                        "details": {
                          "items": [
                            {
                              "product_id": "p_lamp",
                              "requested": 4,
                              "available": 3
                            }
                          ]
                        }
                      }
                    }
                  },
                  "CART_ALREADY_CHECKED_OUT": {
                    "summary": "Cart is frozen after checkout",
                    "value": {
                      "error": {
                        "code": "CART_ALREADY_CHECKED_OUT",
                        "message": "Cart has already been checked out",
                        "details": {
                          "order_id": "bb5df4bff6d1471cb976cb781c95a385"
                        }
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "Quantity out of range",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Request body is invalid",
                        "details": {
                          "errors": [
                            {
                              "type": "greater_than",
                              "loc": [
                                "body",
                                "quantity"
                              ],
                              "msg": "Input should be greater than 0",
                              "input": 0,
                              "ctx": {
                                "gt": 0
                              }
                            }
                          ]
                        }
                      }
                    }
                  },
                  "VALIDATION_ERROR_EXTRA": {
                    "summary": "Unknown field in the body",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Request body is invalid",
                        "details": {
                          "errors": [
                            {
                              "type": "extra_forbidden",
                              "loc": [
                                "body",
                                "coupon"
                              ],
                              "msg": "Extra inputs are not permitted",
                              "input": "SAVE-7F3A9C21"
                            }
                          ]
                        }
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{order_id}": {
      "get": {
        "tags": [
          "Checkout & orders"
        ],
        "summary": "Get an order",
        "description": "Orders never change after they're placed. Names and prices are copies from checkout time,\nso they still explain the total after the catalogue changes.",
        "operationId": "get_order_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Order ID.",
              "examples": [
                "bb5df4bff6d1471cb976cb781c95a385"
              ],
              "title": "Order Id"
            },
            "description": "Order ID."
          }
        ],
        "responses": {
          "200": {
            "description": "The order exactly as placed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "examples": {
                  "coupon": {
                    "summary": "Order with a coupon",
                    "value": {
                      "id": "bb5df4bff6d1471cb976cb781c95a385",
                      "cart_id": "b685d68c87cb40ffbb3c88af5c8cdfad",
                      "subtotal_minor": 10997,
                      "coupon_code": "SAVE-7F3A9C21",
                      "discount_percent": 10,
                      "discount_minor": 1100,
                      "total_minor": 9897,
                      "currency": "INR",
                      "created_at": "2026-10-06T18:26:10.303402Z",
                      "lines": [
                        {
                          "product_id": "p_cable",
                          "product_name": "USB-C Cable",
                          "unit_price_minor": 999,
                          "quantity": 2,
                          "line_total_minor": 1998
                        },
                        {
                          "product_id": "p_keyboard",
                          "product_name": "Mechanical Keyboard",
                          "unit_price_minor": 8999,
                          "quantity": 1,
                          "line_total_minor": 8999
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "examples": {
                  "ORDER_NOT_FOUND": {
                    "summary": "Unknown order ID",
                    "value": {
                      "error": {
                        "code": "ORDER_NOT_FOUND",
                        "message": "Order 7c1e9a3f5b2d4e6a8c0b1d3f5e7a9c2b does not exist",
                        "details": {}
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/coupons": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List coupons",
        "operationId": "list_coupons_admin_coupons_get",
        "responses": {
          "200": {
            "description": "All coupons, oldest milestone first.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Coupon"
                  },
                  "type": "array",
                  "title": "Response List Coupons Admin Coupons Get"
                },
                "examples": {
                  "coupons": {
                    "summary": "One redeemed, one available",
                    "value": [
                      {
                        "code": "SAVE-7F3A9C21",
                        "percent": 10,
                        "milestone_order": 5,
                        "created_at": "2026-10-06T18:26:10.303402Z",
                        "redeemed_order_id": "bb5df4bff6d1471cb976cb781c95a385",
                        "status": "redeemed"
                      },
                      {
                        "code": "SAVE-2B8E40D6",
                        "percent": 10,
                        "milestone_order": 10,
                        "created_at": "2026-10-06T19:02:44.918204Z",
                        "status": "available"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Generate a coupon",
        "description": "Issues **one** coupon for the oldest milestone that has been reached but not rewarded yet.\nWith n = 5, milestones fall at orders 5, 10, 15, and so on; every placed order counts, including discounted ones.\nIf several milestones are owed, call again for each. Concurrent calls can never issue two coupons for the same milestone.\n\nThe coupon gets the current x%, which stays fixed even if the configuration changes later. After a change to n,\nthe next milestone is `last rewarded + n`.",
        "operationId": "generate_coupon_admin_coupons_post",
        "responses": {
          "201": {
            "description": "Coupon issued for the oldest unrewarded milestone.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coupon"
                },
                "examples": {
                  "issued": {
                    "summary": "Coupon for order milestone 5",
                    "value": {
                      "code": "SAVE-7F3A9C21",
                      "percent": 10,
                      "milestone_order": 5,
                      "created_at": "2026-10-06T18:26:10.303402Z",
                      "status": "available"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict with current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "NO_ELIGIBLE_MILESTONE": {
                    "summary": "No unrewarded milestone reached yet",
                    "value": {
                      "error": {
                        "code": "NO_ELIGIBLE_MILESTONE",
                        "message": "No unrewarded order milestone has been reached",
                        "details": {
                          "orders_placed": 7,
                          "next_milestone_order": 10
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id",
            "description": "ID of an existing product, e.g. `p_keyboard`."
          },
          "quantity": {
            "type": "integer",
            "maximum": 1000.0,
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Whole units, 1\u20131000. Must be a JSON integer: `1.5` and `\"2\"` are rejected."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "AddItem"
      },
      "Cart": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "checked_out"
            ],
            "title": "Status"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id",
            "description": "The order this cart became, once checked out."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartItem"
            },
            "type": "array",
            "title": "Items"
          },
          "subtotal_minor": {
            "type": "integer",
            "title": "Subtotal Minor",
            "description": "Sum of line totals at current prices, before any coupon."
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "ISO 4217 code all amounts on this object are in, e.g. `INR`. Amounts are integers in its minor unit: paise for INR, cents for USD, yen for JPY."
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "order_id",
          "items",
          "subtotal_minor",
          "currency"
        ],
        "title": "Cart"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable, machine-readable error code. Branch on this, not on `message`."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Structured data the client can act on; `{}` when there is none."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
    }
  },
...
}